% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_init_k.R
\name{arg_init_k}
\alias{arg_init_k}
\title{Accepted values of argument init_k}
\usage{
arg_init_k()
}
\description{
Returns posssible value of init_k (initialisation value of K for model
optimisation).
}
\details{
\code{true_k} refers to optimisation results initialised with K set to
the true value (40), i.e. results produced with \code{\link[=run_optim]{run_optim()}}
\code{from_n} refers to optimisation results initialised with K set to
the nb of tips in a tree, i.e. results produced with \code{\link[=run_optim_from_n]{run_optim_from_n()}}
}
\author{
Theo Pannetier
}
