% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_optim.R
\name{arg_optim}
\alias{arg_optim}
\title{Possible values of argument optim}
\usage{
arg_optim()
}
\description{
Names of the diversification models currently accepted as arguments by
various functions of \code{DDvTDtools}.
}
\author{
Theo Pannetier
}
