% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_sim.R
\name{arg_sim}
\alias{arg_sim}
\title{Possible values of argument sim}
\usage{
arg_sim()
}
\description{
Names of the simulation models currently accepted as arguments by various
functions of \code{DDvTDtools}.
}
\details{
DD refers to the diversity-dependent model, TD to the time-dependent
model.
}
\author{
Theo Pannetier
}
