% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_landscape_data.R
\name{assemble_landscape_data}
\alias{assemble_landscape_data}
\title{Assemble log-likelihood landscape data for plotting}
\usage{
assemble_landscape_data(sim, para, mc)
}
\arguments{
\item{sim}{character, the name of a simulation model. See
\code{get_sim_names()} for possible values.}

\item{para}{numeric or character. A four-digits code specifying a set of
parameter values.}

\item{mc}{numeric, the index of the tree to get the landscape for.}
}
\description{
Fetches saved results for likelihood landscapes along K and assembles a data
frame ready for plotting
}
\author{
Théo Pannetier
}
