% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_DDvTD_wd.R
\name{assert_DDvTD_wd}
\alias{assert_DDvTD_wd}
\title{Assert that DDvTD is the working directory}
\usage{
assert_DDvTD_wd()
}
\description{
Many functions in DDvTDtools rely on a relative path from DDvTD folder to
access data and save files. This function asserts R is set at this folder
before proceeding.
}
\author{
Théo Pannetier
}
