% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_init.R
\name{assert_init}
\alias{assert_init}
\title{Asserts input init parameter is supported}
\usage{
assert_init(init)
}
\arguments{
\item{init}{a character for the initial parameter setting, typically passed
from another function}
}
\description{
Asserts input init parameter is supported
}
\author{
Theo Pannetier
}
