% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_init_k.R
\name{assert_init_k}
\alias{assert_init_k}
\title{Asserts input init parameter is supported}
\usage{
assert_init_k(init_k)
}
\arguments{
\item{init_k}{a character for the initial parameter setting, typically passed
from another function}
}
\description{
Asserts input init parameter is supported
}
\author{
Theo Pannetier
}
