% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_sim.R
\name{assert_sim}
\alias{assert_sim}
\title{Assert the value of 'sim' is a valid model name for DDvTDtools}
\usage{
assert_sim(sim)
}
\arguments{
\item{sim}{character. The name of a simulation model}
}
\description{
Assert the value of 'sim' is a valid model name for DDvTDtools
}
\author{
Theo Pannetier
}
