% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convergence_optim.R
\name{check_convergence_optim}
\alias{check_convergence_optim}
\title{Check optimisation results for completeness}
\usage{
check_convergence_optim(folder = "data/optim/")
}
\arguments{
\item{folder}{the folder where the results files are stored. All .rds files in the folder are read.}
}
\description{
The function reads results files output from run_ML calls and prints whether some
results are missing or some optimisations have not converged
}
\author{
Théo Pannetier
}
