% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_results.R
\name{check_duplicated_rows}
\alias{check_duplicated_rows}
\title{Check a resultsults data frame does not have duplicated rows}
\usage{
check_duplicated_rows(results, return_rows)
}
\arguments{
\item{results}{a resultsults data frame, as produced by [run_optim()] or accessed by
[read_optim_resultsults].}

\item{return_rows}{logical. Should rows that don't pass the check be
returned?}
}
\description{
Goes through rows and returns a message if two rows are found to be
duplicates.
}
\author{
Theo Pannetier
}
