% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_results.R
\name{check_incomplete_rows}
\alias{check_incomplete_rows}
\title{Check a resultsults data frame for NAs in rows}
\usage{
check_incomplete_rows(results, return_rows)
}
\arguments{
\item{results}{a resultsults data frame, as produced by [run_optim()] or accessed by
[read_optim_resultsults].}

\item{return_rows}{logical. Should rows that don't pass the check be
returned?}
}
\description{
Goes through rows and returns a message if any NA is found in  row.
}
\author{
Theo Pannetier
}
