% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_init_pars.R
\name{check_init_pars}
\alias{check_init_pars}
\title{Check initial parameter values for run_optim}
\usage{
check_init_pars(init_pars, optim, brts)
}
\arguments{
\item{init_pars}{numeric vector. The initial parameter values to be checked.}

\item{optim}{character, name of the optimisation model, either \code{"DD"}
or \code{"TD"}.}

\item{brts}{numeric vector, a set of branching times as given by
\code{ape::branching_times()}.}
}
\description{
Checks that default or user-specified values entered to initialize the
ML optimisation do not break forbidden conditions inside \code{DDD::dd_ML()}
and \code{DDD::bd_ML()}.
}
\author{
Théo Pannetier
}
