% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_initpars.R
\name{check_initpars}
\alias{check_initpars}
\title{Check initial parameter values for run_ML}
\usage{
check_initpars(init_pars, optim, brts)
}
\description{
Internal function called by \code{run_ML}. The function checks that default
or user-specified values entered to initialize the ML optimisation do not break
forbidden conditions inside \code{DDD::dd_ML()} and \code{DDD::bd_ML()}.
}
\author{
Théo Pannetier
}
