% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_results.R
\name{check_loglikelihood_exists}
\alias{check_loglikelihood_exists}
\title{Check a resultsults data frame for log-likelihood}
\usage{
check_loglikelihood_exists(results, return_rows)
}
\arguments{
\item{results}{a resultsults data frame, as produced by [run_optim()] or accessed by
[read_optim_resultsults].}

\item{return_rows}{logical. Should rows that don't pass the check be
returned?}
}
\description{
Goes through rows and returns a message if a row lacks a valid
log- likelihood value
}
\author{
Theo Pannetier
}
