% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_results.R
\name{check_missing_mcs}
\alias{check_missing_mcs}
\title{Check a results data frame for missing rows}
\usage{
check_missing_mcs(results, return_rows)
}
\arguments{
\item{results}{a results data frame, as produced by \code{\link[=run_optim]{run_optim()}} or accessed by
\code{\link[=read_optim_results]{read_optim_results()}}.}

\item{return_rows}{logical. Should rows that don't pass the check be
returned?}
}
\description{
Goes through rows and returns a message if any row from 1 to 1000 is missing
}
\author{
Theo Pannetier
}
