% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_bd_ML_version.R
\name{compare_bd_ML_version}
\alias{compare_bd_ML_version}
\title{Compare likelihood values outputed by two versions of bd_ML() using DDvTD results}
\usage{
compare_bd_ML_version(sim, para)
}
\arguments{
\item{sim}{name of the model used to simulate the trees. See \code{get_sim_names()}}

\item{para}{four-digit-character specifying the parameters used to simulate the trees. See \code{get_para_values}.}
}
\description{
Assembles a data frame containing ML estimates from bd_ML, bd_ML_3.8, and previous results from DDvTD.
}
\details{
\code{DDvTDtools} parameter \code{init} is set to 1 (true value). The comparison is run on the 50 first trees.
}
\author{
Theo Pannetier
}
