% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_dd_loglik_version.R
\name{compare_dd_loglik_version}
\alias{compare_dd_loglik_version}
\title{Compare likelihood values outputed by two versions of dd_loglik() using DDvTD results}
\usage{
compare_dd_loglik_version(sim, para, methode = "analytical",
  sim_dir = "./data/sim/", optim_dir = "./data/optim/",
  output_dir = "./DDD_loglik_version_comparison/")
}
\arguments{
\item{sim}{name of the model used to simulate the trees. See \code{get_sim_names()}}

\item{para}{four-digit-character specifying the parameters used to simulate the trees. See \code{get_para_values}.}

\item{methode}{character, optimization methode passed to \code{dd_loglik()}. See \code{DDD} documentation for details.}

\item{sim_dir}{character, path to the directory where the sim file is located.}

\item{optim_dir}{character, path to the directory where the optim file is located.}

\item{output_dir}{character, path to the directory where to save the output.}
}
\description{
Reads parameter values in an 'optim" dataset (see DDvTD documentation) and calls \code{dd_loglik()} from \code{DDD.3.2} and \code{3.8}
}
\details{
\code{DDvTDtools} parameter \code{init} is set to 1 (true value).
}
\author{
Theo Pannetier
}
