% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_method_dd.R
\name{compare_method_dd}
\alias{compare_method_dd}
\title{Peform a comparison of integration methods for DDD::dd_ML}
\usage{
compare_method_dd(sim, para)
}
\arguments{
\item{sim}{either "DD" or "TD", the simulation model.}

\item{para}{the parameters code.}
}
\description{
This function runs dd_ML() three times for the 50 first trees in the dataset selected by the arguments.
Each time a different \code{method} argument is used, among \code{"lsoda"}, \code{"ode45"}, \code{"analytical"}.
}
\author{
Théo Pannetier
}
