% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_errors.R
\name{count_errors}
\alias{count_errors}
\title{Count errors in a log file}
\usage{
count_errors(log_filename, error_templates)
}
\arguments{
\item{error_templates}{a character vector with the error patterns to look 
for.
The match must be exact for an error to be counted.}

\item{log_file}{a log file to search the errors in. \code{.txt} or 
\code{.log} files work, maybe other formats?}
}
\value{
a vector of the same length as \code{error_templates} giving 
the number of counts of each error.
}
\description{
Counts the number of lines in a text file matching specified errors.
}
\author{
Théo Pannetier
}
