% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_errors_logs.R
\name{count_errors_log}
\alias{count_errors_log}
\title{Count errors in a log file}
\usage{
count_errors_log(path_to_log)
}
\arguments{
\item{path_to_log}{path to a log file to search the errors in.
The log file can be a \code{.txt} or \code{.log} file.}
}
\value{
a list with the indexes of trees that returned each error.
Also prints the nb of occurrences of each error to the console
}
\description{
Reports occcurrences of common errors in log files generated by
\code{run_optim()}.
}
\author{
Théo Pannetier
}
