% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cure_error_rows.R
\name{cure_error_rows}
\alias{cure_error_rows}
\title{Cure format for erroneous results}
\usage{
cure_error_rows(df)
}
\arguments{
\item{df}{a results data frame, i.e. as read by
\code{DDvTDtools::read_optim_table()}}
}
\description{
Changes the format of results for trees on which optimisation failed, e.g.
changing loglik = -1 to loglik = -Inf.
}
\author{
Theo Pannetier
}
