% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDvTD_labels.R
\docType{data}
\name{DDvTD_tags}
\alias{DDvTD_tags}
\title{Models currently accepted in DDvTDtools}
\format{A character vector of any length.}
\usage{
data(DDvTD_tags)
}
\description{
A character vector containing tags for all models currently accepted by DDvTDtools.
Tags should be very short, 2-4 characters, e.g. \code{"DD"} for diversity-dependent
diversification or \code{"CR"} for constant-rate diversification.
}
\keyword{datasets}
