% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_pars2_dd1.R
\name{default_pars2_dd1}
\alias{default_pars2_dd1}
\title{Default DD pars2 argument passed to DDD}
\usage{
default_pars2_dd1()
}
\description{
Default values for the \code{pars2} argument of \code{\link[DDD:dd_loglik]{DDD::dd_loglik()}}
}
\author{
Theo Pannetier
}
