% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_pars2_td4.R
\name{default_pars2_td4}
\alias{default_pars2_td4}
\title{Default TD pars2 argument passed to DDD}
\usage{
default_pars2_td4()
}
\description{
Default values for the \code{pars2} argument of \code{\link[DDD:bd_loglik]{DDD::bd_loglik()}}
}
\author{
Theo Pannetier
}
