% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_DDvTD_tags.R
\name{edit_DDvTD_tags}
\alias{edit_DDvTD_tags}
\title{Edit the model tags accepted by DDvTDtools}
\usage{
edit_DDvTD_tags(tags)
}
\arguments{
\item{tags}{a character vector containing **all** the model tags to be accepted (does not append).}
}
\value{
None.
}
\description{
Overwrites the vector \code{DDvTD_labels} in \code{/data}. Tags should be very short, 2-4 characters,
e.g. \code{"DD"} for diversity-dependent diversification or \code{"CR"} for constant-rate diversification.
}
\author{
Theo Pannetier, \email{t.s.c.pannetier@rug.nl}
}
