% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_rows.R
\name{fill_missing_rows}
\alias{fill_missing_rows}
\title{Fill missing results row with default values}
\usage{
fill_missing_rows(sim, optim, para)
}
\arguments{
\item{sim}{character. The model used to simulate the tree.}

\item{optim}{character. The model fitted to the tree.}

\item{para}{numeric or character. A four-digits code specifying a set of
parameter values. See \code{get_para_values()} for possible values.}
}
\description{
Function used to fill result data for trees that miss any result.
Fills the result matrix for a parameter set with default values for these
missing trees
}
\author{
Théo Pannetier
}
