% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_td_setophaga.R
\name{fit_td_setophaga}
\alias{fit_td_setophaga}
\title{Fit the time-dependent model on Setophaga phylogeny}
\usage{
fit_td_setophaga(lambda_ini, mu_ini, k_ini, outputfile)
}
\arguments{
\item{lambda_ini}{numeric, the initial value for parameter lambda0}

\item{mu_ini}{numeric, the initial value for parameter mu0}

\item{k_ini}{numeric, the initial value for parameter K}

\item{outputfile}{character. Name of the outputfile should be a .rds}
}
\description{
Run the maximum likelihood of the time-dependent model on the phylogeny of
Setophaga warblers
}
\author{
Théo Pannetier
}
