% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_indexing.R
\name{fix_indexing}
\alias{fix_indexing}
\title{Fix indexing of results data frame}
\usage{
fix_indexing(df)
}
\arguments{
\item{df}{a results data frame, i.e. as read by
\code{DDvTDtools::read_optim_table()}}
}
\description{
Checks if the row are indexed correctly (from 1 to 1000), and, if not, re-
orders them accordingly.
}
\author{
Theo Pannetier
}
