% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_optim_results_row.R
\name{format_optim_results_row}
\alias{format_optim_results_row}
\title{Assemble a row of results for run_optim output}
\usage{
format_optim_results_row(
  mc,
  sim,
  optim,
  brts,
  true_pars,
  init_pars,
  ML_output,
  num_cycles = 1,
  methode,
  optimmethod,
  jobID,
  cond
)
}
\arguments{
\item{mc}{numeric between 1 and 1000, the index of a tree within a dataset.}

\item{sim}{character, name of simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{optim}{character, name of the optimisation model, either \code{"DD"}
or \code{"TD"}.}

\item{brts}{numeric vector, a set of branching times as given by
\code{ape::branching_times()}.}

\item{true_pars}{numeric vector. The true parameter values used to simulate
the tree.}

\item{init_pars}{numeric vector. The initial parameter values used for
likelihood optimisation.}

\item{ML_output}{numeric vector. Raw output of \code{\link[DDD:dd_ML]{DDD::dd_ML()}} /
\code{\link[DDD:bd_ML]{DDD::bd_ML()}}.}

\item{num_cycles}{numeric. Maximum number of optimisation cycles.}

\item{methode}{character. Integration methode for computing the likelihood.
See \code{\link[DDD:dd_loglik]{DDD::dd_loglik()}} / \code{\link[DDD:bd_loglik]{DDD::bd_loglik()}} for details.}

\item{optimmethod}{character. Optimisation methode used to compute the
maximum likelihood. See \code{\link[DDD:dd_ML]{DDD::dd_ML()}} / \code{\link[DDD:bd_ML]{DDD::bd_ML()}} for details.}

\item{jobID}{the ID number of the cluster job that returned the results.
Saved in metadata to retrieve logs, for example. A value of \code{NULL} means
that the result was obtained locally.}

\item{cond}{numeric, conditionning parameter passed to
\code{\link[DDD:dd_ML]{DDD::dd_ML()}} / \code{\link[DDD:bd_ML]{DDD::bd_ML()}}.}
}
\description{
Internal function called by \code{\link[=run_optim]{run_optim()}}. Formats optimisation results and
metadata in a data frame row.
}
\author{
Théo Pannetier
}
