% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brts.R
\name{get_brts}
\alias{get_brts}
\title{Extract a set of branching times from a dataset of simulated trees}
\usage{
get_brts(sim, para, mc, with_extinct = F)
}
\arguments{
\item{sim}{character, name of simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{get_para_values} doc for details, and call
\code{get_para_values} to see possible inputs.}

\item{mc}{numeric between 1 and 1000, the index of a tree within a dataset.}

\item{with_extinct}{logical, should the tree include extinct lineages?}
}
\value{
a numerical vector containing the branching times of the tree.
}
\description{
Read the dataset and returns the branching times for the specified tree.
}
\author{
Theo Pannetier
}
