% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dAIC_DDvTD.R
\name{get_dAIC_DDvTD}
\alias{get_dAIC_DDvTD}
\title{Compute AIC_DD - AIC_TD for a set of trees}
\usage{
get_dAIC_DDvTD(sim, para)
}
\description{
Compute AIC_DD - AIC_TD for a set of trees
}
\author{
Théo Pannetier
}
