% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_initpars.R
\name{get_default_initpars}
\alias{get_default_initpars}
\title{Get default initial parameter values for run_ML optimisation}
\usage{
get_default_initpars(true_pars, optim, brts)
}
\description{
Internal function called only by \code{run_ML}.
}
\author{
Théo Pannetier
}
