% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_delta_aic.R
\name{get_delta_aic}
\alias{get_delta_aic}
\title{Computes the AIC difference between DD and TD}
\usage{
get_delta_aic(sim, para, init_k = "true_k")
}
\arguments{
\item{sim}{character, name of simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{arg_para()} doc for details, and call
\code{arg_para()} to see possible inputs.}

\item{init_k}{character code specifying the initial values of parameter K
used in the optimisation. Refer to \code{get_possible init_k()} doc for
details, and call \code{arg_init_k()} to see possible inputs.}
}
\description{
Reads a pair of \code{optim = "DD"} and \code{optim = "TD"} results data frames and
computes the AIC difference between them.
}
\author{
Théo Pannetier
}
