% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_empty_optim_df.R
\name{get_empty_optim_df}
\alias{get_empty_optim_df}
\title{Get an empty skeleton for run_ML output data frame.}
\usage{
get_empty_optim_df()
}
\description{
Internal function called only by \code{run_ML}.
}
\author{
Théo Pannetier
}
