% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_init_k.R
\name{get_init_k}
\alias{get_init_k}
\title{Get values for argument init_k}
\usage{
get_init_k(incl_unused = FALSE)
}
\arguments{
\item{incl_unused}{logical, specify whether you want to include values that
have been considered in earlier phases of the project.}
}
\description{
Labels for initial value of K associated with each parameter setting for the
main analysis.
}
\author{
Théo Pannetier
}
