% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Kprime.R
\name{get_Kprime}
\alias{get_Kprime}
\title{Compute K' from la0, mu0, K}
\usage{
get_Kprime(pars)
}
\arguments{
\item{pars}{a numeric vector containing c(la0, mu0, K)}
}
\description{
Compute K' from la0, mu0, K
}
\author{
Théo Pannetier
}
