% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lr.R
\name{get_lr}
\alias{get_lr}
\title{Compute likelihood ratios for a results dataset}
\usage{
get_lr(sim, para, init_k = "true_k")
}
\arguments{
\item{sim}{character. The name of a simulation model}

\item{para}{numeric or character. A four-digits code specifying a set of
parameter values. See \code{get_para_values()} for possible values.}

\item{init_k}{character, the setting used to initialise parameter K.
See \code{get_possible_init_k()} for possible values.}
}
\description{
Compute likelihood ratios for a results dataset
}
\author{
Théo Pannetier
}
