% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lr_table.R
\name{get_lr_table}
\alias{get_lr_table}
\title{Assemble a table containing likelihood ratios}
\usage{
get_lr_table(para, init_k)
}
\arguments{
\item{init_k}{}
}
\description{
Calls \code{get_lr} and binds results in a table with corresponding
\code{sim} ans \code{para} values.
}
\author{
Theo Pannetier
}
