% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lr_tbl.R
\name{get_lr_tbl}
\alias{get_lr_tbl}
\title{Assemble a table containing likelihood ratios}
\usage{
get_lr_tbl(
  sim,
  para,
  init_k = get_init_k()[which(names(get_init_k()) == para)]
)
}
\arguments{
\item{sim}{character, name of the simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits number coding for a set of four parameter
values. Refer to \code{arg_para()} doc for details, and call
\code{arg_para()} to see possible inputs.}

\item{init_k}{character code specifying the initial values of parameter K
used in the optimisation. Refer to \code{arg_init_k()} doc for
details, and call \code{arg_init_k()} to see possible inputs.}
}
\description{
Calls \code{\link[=get_lr_tbl]{get_lr_tbl()}} for a set of simulations and returns the
log-likelihood ratio in a table.
}
\author{
Theo Pannetier
}
