% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ltt_tbls.R
\name{get_ltt_tbls}
\alias{get_ltt_tbls}
\title{Get LTT tables and average LTT for a set of trees}
\usage{
get_ltt_tbls(sim, para, with_extinct = FALSE)
}
\arguments{
\item{sim}{character, name of the simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits number coding for a set of four parameter
values. Refer to \code{arg_para()} doc for details, and call
\code{arg_para()} to see possible inputs.}

\item{with_extinct}{logical, should the tree include extinct lineages?}
}
\value{
a list of tibbles of the LTT coordinates.
The first element corresponds to the average LTT, and each other element
is to the LTT of a single tree.
}
\description{
Calls \code{\link[TreeSim:LTT.plot.gen]{TreeSim::LTT.plot.gen()}} to get LTT coordinates of trees in a
simulated data set, and the average LTT coordinates
}
\author{
Théo Pannetier
}
