% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multi_brts.R
\name{get_multi_brts}
\alias{get_multi_brts}
\title{Extract multiple sets of branching times from a dataset of simulated trees}
\usage{
get_multi_brts(sim, para, rangemc = NULL, with_extinct = FALSE)
}
\arguments{
\item{sim}{character, name of simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{arg_para()} doc for details, and call
\code{arg_para()} to see possible inputs.}

\item{rangemc}{numeric vector, a set of tree indices ranging from 1 to 1000.}

\item{with_extinct}{logical, should the tree include extinct lineages?}
}
\value{
a list of numerical vectors containing the branching times of the specified trees.
}
\description{
Read the dataset at specified location and returns branching times for
multiple trees specified by indices.
}
\author{
Theo Pannetier
}
