% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_table.R
\name{get_n_table}
\alias{get_n_table}
\title{Get N for each tree in a dataset}
\usage{
get_n_table(sim, para)
}
\arguments{
\item{sim}{character, name of simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{get_para_values} doc for details, and call
\code{get_para_values} to see possible inputs.}
}
\description{
Assemble a table containing the number of tips in each tree in a results data
set.
}
\author{
Theo Pannetier
}
