% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optim_names.R
\name{get_optim_names}
\alias{get_optim_names}
\title{Get possible values of optim argument}
\usage{
get_optim_names()
}
\description{
Names of the diversification models currently accepted as arguments by various functions of \code{DDvTDtools}.
}
\author{
Theo Pannetier
}
