% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_DD.R
\name{get_p_DD}
\alias{get_p_DD}
\title{Compute the proportion of trees for which DD is preferred}
\usage{
get_p_DD(sim, para, init_k = "true_k")
}
\arguments{
\item{sim}{character. The name of a simulation model}

\item{para}{numeric or character. A four-digits code specifying a set of
parameter values. See \code{get_para_values()} for possible values.}

\item{init_k}{character, the setting used to initialise parameter K.
See \code{get_possible_init_k()} for possible values.}
}
\description{
Returns p_DD, the proportion of trees in a parameter setting for which DD
fits better than TD (i.e. log-likelihood ratio > 0)
}
\author{
Théo Pannetier
}
