% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_DD_table.R
\name{get_p_DD_table}
\alias{get_p_DD_table}
\title{Get p_DD values for each model and parameter setting}
\usage{
get_p_DD_table()
}
\description{
Returns values of p_DD (see \code{get_p_DD}) for both models and all parameter
settings used in the main analysis
}
\author{
Théo Pannetier
}
