% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_success.R
\name{get_p_success}
\alias{get_p_success}
\title{Computes the proportion of trees that can successfully be assigned to either
model}
\usage{
get_p_success(sim, para, init_k)
}
\arguments{
\item{init_k}{}
}
\description{
Computes the proportion of trees in the dataset that reject the opposite
model, i.e. logLR is beyond the corresponding threshold computed by
\code{get_lr_threshold}.
}
\author{
Theo Pannetier
}
