% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_tbl.R
\name{get_p_tbl}
\alias{get_p_tbl}
\title{Compute threshold values and proportions of correct model inference}
\usage{
get_p_tbl(paras)
}
\arguments{
\item{paras}{numeric or character vector, a set of \code{para} values. See
\link{arg_para} for possible values.}
}
\value{
a \code{tibble} with the parameter values corresponding to \code{para}, the
threshold values based on percentiles, and the proportion of trees beyond
the threshold for each model.
}
\description{
Given a set of parameters, computes the 5th and 95th percentile of the
log-likelihood ratio for DD trees and TD trees, respectively, and the
proportion of trees of the other distribution with log-likelihood ratio
scores beyond those thresholds; i.e., the probability that the right model
is recovered from such a tree.
}
\author{
Théo Pannetier
}
