% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_para_values.R
\name{get_para_values}
\alias{get_para_values}
\title{Get possible values of para argument}
\usage{
get_para_values(incl_unused = FALSE)
}
\arguments{
\item{incl_unused}{logical, specify whether you want to include values that
have been considered in earlier phases of the project.}
}
\description{
Input values currently accepted for para.
   Use \code{para_to_pars} to translate para into paramter values.
}
\seealso{
\code{para_to_pars()}
}
\author{
Theo Pannetier
}
