% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_possible_init_k.R
\name{get_possible_init_k}
\alias{get_possible_init_k}
\title{Accepted values of argument init_k}
\usage{
get_possible_init_k()
}
\description{
Returns posssible value of init_k (initialisation value of K for model
optimisation). For \code{true_k} K is initialised at the true value (40),
for \code{from_n} it is initialised at N, the nb  of tips in the tree
}
\author{
Theo Pannetier
}
