% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_possible_inits.R
\name{get_possible_inits}
\alias{get_possible_inits}
\title{Accepted values of argument init}
\usage{
get_possible_inits()
}
\description{
Returns posssible value of init (initialisation setting). For \code{true_k}
K is initialised at the true value (40), for \code{from_n} it is initialised
at N, the nb  of tips in the tree
}
\author{
Theo Pannetier
}
