% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_names.R
\name{get_sim_names}
\alias{get_sim_names}
\title{Get possible values of sim argument}
\usage{
get_sim_names()
}
\description{
Names of the simulation models currently accepted as arguments by various functions of \code{DDvTDtools}. Based on \code{becosys::get_bd_param_names} from Richel J.C. Bilderbeek.
}
\author{
Theo Pannetier
}
