% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_ltt_tbl.R
\name{get_summary_ltt_tbl}
\alias{get_summary_ltt_tbl}
\title{Compute summary metrics for a set of phylogenetic trees}
\description{
See \code{magrittr::\link[phylobates:get_summary_ltt_tbl]{\%>\%}} for details.
}
\keyword{internal}
