% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_tags.R
\docType{data}
\name{init_tags}
\alias{init_tags}
\title{Initial conditions currently accepted in DDvTDtools}
\format{A numerical vector which elements are a single-digit numerical value.}
\usage{
data("init_tags")
}
\description{
A vector containing the currently accepted initial conditions in DDvTDtools.
}
\keyword{datasets}
