% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_para_accepted.R
\name{is_para_accepted}
\alias{is_para_accepted}
\title{Is the value of 'para' a valid parameter combination for DDvTDtools?}
\usage{
is_para_accepted(para)
}
\arguments{
\item{para}{numeric or character. A four-digits code specifying a set of parameter values.}
}
\description{
Is the value of 'para' a valid parameter combination for DDvTDtools?
}
\seealso{
\code{get_para_values()}, \code{para_to_pars()}
}
\author{
Theo Pannetier
}
