% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sim_accepted.R
\name{is_sim_accepted}
\alias{is_sim_accepted}
\title{Is the value of 'sim' a valid model name for DDvTDtools?}
\usage{
is_sim_accepted(sim)
}
\arguments{
\item{sim}{character. The name of a simulation model}
}
\description{
Is the value of 'sim' a valid model name for DDvTDtools?
}
\author{
Theo Pannetier
}
