% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_optim_tbls.R
\name{join_optim_tbls}
\alias{join_optim_tbls}
\title{Read a pair of \code{optim} tables and join them together}
\usage{
join_optim_tbls(
  sim,
  para,
  init_k = get_init_k()[which(names(get_init_k()) == para)]
)
}
\arguments{
\item{sim}{character, name of the simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits number coding for a set of four parameter
values. Refer to \code{arg_para()} doc for details, and call
\code{arg_para()} to see possible inputs.}

\item{init_k}{character code specifying the initial values of parameter K
used in the optimisation. Refer to \code{arg_init_k()} doc for
details, and call \code{arg_init_k()} to see possible inputs.}
}
\description{
For a given combination of \code{sim} and \code{para}, this function will fetch the
two corresponding optimisation results tables; filter out the metadata
and bind them together. Also computes the log likelihood ratio along the way.
}
\author{
Théo Pannetier
}
