% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_model.R
\name{optimize_model}
\alias{optimize_model}
\title{Optimize a diversification model on a set of simulated phylogenetic trees.}
\usage{
optimize_model(sim_model, para, optim_model, init = 1,
  inputfile = paste0("./data/sim/", "sim", sim_model, "-", para,
  ".RData"), outputfile = paste0("./data/optim/", "sim", sim_model,
  "_optim", optim_model, "_init", init, "-", para, ".rds"),
  rangemc = NULL, overwrite = F, ...)
}
\arguments{
\item{sim_model}{a character used to identify the desired simulated input dataset. See \code{DDvTD_tags} for a list of possible values.}

\item{para}{a 4-digit code used to identify the desired simulated input dataset.}

\item{optim_model}{a character specifying the model to be optimized on input data. See \code{DDvTD_tags} for a list of possible values.}

\item{init}{an integer controlling the initial parameter values to be used.}

\item{inputfile}{path and name of the input file, by default generated automatically from \code{sim_model} and \code{para}}

\item{outputfile}{path and name for output file, by default generated automatically from \code{sim_model}, \code{para}, \code{optim_model} and \code{init}.}

\item{rangemc}{a numeric vector containing all the indices of the trees to optimize the model on. Default to all the trees in the dataset.}

\item{overwrite}{logical. \code{optimize_model} always try to load previous results if they exist. If \code{overwrite} is \code{FALSE} previous results will be kept and trees for which results already exist are excluded from \code{rangemc}.}

\item{...}{additional parameter values to be passed to \code{dd_ML} or \code{bd_ML}, e.g. \code{cond}, \code{tol} or \code{methode}.}
}
\value{
A \code{data.frame} including results and metadata:
\itemize{
   \item{ \code{sim}} the model used to generate simulated data
   \item{ \code{true_lambda0}} parameter value used to generate simulated data
   \item{ \code{true_mu0}} parameter value used to generate simulated data
   \item{ \code{true_K}} parameter used to generate simulated data
   \item{ \code{mc}} tree index in the dataset
   \item{ \code{df}} number of parameters estimated in the optimization of the model
   \item{ \code{optim}} optimized model
   \item{ \code{init}} initial parameter values setting
   \item{ \code{conv}} exit status of the optimization. conv = 0 means convergence.
   \item{ \code{loglik}} log-likelihood estimate associated with optimized parameter values below
   \item{ \code{AIC}} AIC value calculated from \code{loglik} and \code{df}
   \item{ \code{lambda0}} maximum likelihood estimate for lambda0
   \item{ \code{mu0}} maximum likelihood estimate for mu0
   \item{ \code{K}} maximum likelihood estimate for K. Takes value \code{Inf} for \code{optim_model = "CR"}.
}
}
\description{
Fit a diversification model on pre-existing simulated phylogenetic trees, and returns a dataset containing the optimized parameter values, maximum likelihood estimate and metadata.
}
\author{
Theo Pannetier, \email{t.s.c.pannetier@rug.nl}
}
