% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para_to_pars.R
\name{para_to_pars}
\alias{para_to_pars}
\title{Translate 'para' argument into parameter values}
\usage{
para_to_pars(para)
}
\arguments{
\item{para}{numeric or character. A four-digits code specifying a set of parameter values. See \code{get_para_values()} for possible values.}
}
\value{
a four-element numeric vector containing the parameter values associated to 'para'.
}
\description{
Prints the parameter values associated with a given value of argument 'para'.
}
\author{
Theo Pannetier
}
