% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_list.R
\docType{data}
\name{parameter_list}
\alias{parameter_list}
\title{Paramter values currently accepted in DDvTDtools}
\format{A list of four numeric vectors of any length.}
\usage{
data("parameter_list")
}
\description{
A lisr that contains the default parameter values of lambda0, mu0 and K
that are implemented in DDvTDtools.
}
\keyword{datasets}
