% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_avg_ltt.R
\name{plot_avg_ltt}
\alias{plot_avg_ltt}
\title{Plot the average LTTs of paired DD and TD trees}
\usage{
plot_avg_ltt(para, with_extinct = FALSE)
}
\arguments{
\item{para}{numeric or character. A four-digits code specifying a set of
parameter values.}

\item{with_extinct}{logical. \code{with_extinct = F} for the reconstructed
tree, \code{with_extinct = T} for the complete tree.}
}
\description{
Simultaneously plots the average ltts of a set of DD and TD trees that share
the same parameter set.
}
\author{
Théo Pannetier
}
