% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_avg_ltt_DDvTD.R
\name{plot_avg_ltt_DDvTD}
\alias{plot_avg_ltt_DDvTD}
\title{Plot the average LTTs of paired DD and TD trees}
\usage{
plot_avg_ltt_DDvTD(para)
}
\arguments{
\item{para}{numeric or character. A four-digits code specifying a set of parameter values.}
}
\description{
Simultaneously plots the average ltts of a set of DD and TD trees that share the same parameter set.
}
\author{
Théo Pannetier
}
