% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_blrt.R
\name{plot_blrt}
\alias{plot_blrt}
\title{Plot logLR distribution for DD & TD trees}
\usage{
plot_blrt(
  para,
  init_k = "true_k",
  plot_thresholds = TRUE,
  label_p_success = FALSE,
  threshold_dd = 0.05,
  threshold_td = 0.95,
  full_range = FALSE,
  which_geom = "density"
)
}
\arguments{
\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{get_para_values} doc for details, and call
\code{get_para_values} to see possible inputs.}

\item{init_k}{character code specifying the initial values of parameter K
used in the optimisation. Refer to \code{get_possible init_k} doc for
details, and call \code{get_possible_init_k} to see possible inputs.}

\item{plot_thresholds}{logical, should quantile bars be plotted with the
distribution?}

\item{label_p_success}{logical, should the proportion of successes be added as
a label on the plot?}

\item{threshold_dd}{numeric, quantiles of the lr distribution that should be
added to the plot.}

\item{threshold_td}{numeric, quantiles of the lr distribution that should be
added to the plot.}

\item{full_range}{logical. If \code{TRUE}, the range of the x-axis is set to
include all points in the dataset. Otherwise, the default is
\code{c(-10, 10)}.}

\item{which_geom}{character giving the type of plot to be drawn. Options are
\code{"density"} and \code{"histograms"}.}
}
\description{
Plot logLR distribution for DD & TD trees
}
\author{
Théo Pannetier
}
