% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dAIC_DDvTD.R
\name{plot_dAIC_DDvTD}
\alias{plot_dAIC_DDvTD}
\title{Plot dAIC distribution for DD & TD trees}
\usage{
plot_dAIC_DDvTD(para)
}
\arguments{
\item{para}{numeric or character. A four-digits code specifying a set of parameter values. See \code{get_para_values()} for possible values.}
}
\description{
Plot dAIC distribution for DD & TD trees
}
\author{
Théo Pannetier
}
