% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dAIC_para.R
\name{plot_dAIC_para}
\alias{plot_dAIC_para}
\title{Plot dAIC distribution for a set of parameter sets}
\usage{
plot_dAIC_para(sim, para_set)
}
\arguments{
\item{sim}{character. The name of a simulation model}

\item{para_set}{numeric or character vector. A set of four-digits codes specifying a set of parameter values.}
}
\description{
Plot the distribution of AIC_TD - AIC_DD for a given \code{sim} and a set of \code{para}.
}
\author{
Théo Pannetier
}
