% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loglik_landscape_K.R
\name{plot_loglik_landscape_K}
\alias{plot_loglik_landscape_K}
\title{Plot likelihood landscapes along K}
\usage{
plot_loglik_landscape_K(sim, para, mc)
}
\arguments{
\item{sim}{character, name of simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{get_para_values} doc for details, and call
\code{get_para_values} to see possible inputs.}

\item{mc}{numeric between 1 and 1000, the index of a tree within a dataset.}
}
\description{
Plot likelihood curves of DD and TD models over parameter K from saved
results.
}
\author{
Théo Pannetier
}
