% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lr.R
\name{plot_lr}
\alias{plot_lr}
\title{Plot log-LR distribution for DD & TD trees}
\usage{
plot_lr(
  para,
  init_k = get_init_k()[which(names(get_init_k()) == para)],
  plot_thresholds = TRUE,
  label_p_success = TRUE,
  quant_range = c(0, 1)
)
}
\arguments{
\item{para}{numeric, a four-digits number coding for a set of four parameter
values. Refer to \code{arg_para()} doc for details, and call
\code{arg_para()} to see possible inputs.}

\item{init_k}{character code specifying the initial values of parameter K
used in the optimisation. Refer to \code{arg_init_k()} doc for
details, and call \code{arg_init_k()} to see possible inputs.}

\item{plot_thresholds}{logical, should threshold bars be plotted with the
distribution?}

\item{label_p_success}{logical, should the proportion of successes be added
as a label on the plot?}

\item{quant_range}{a length-2 numeric vector, the quantiles defining the
range of values to be plotted}
}
\description{
Produces the log-LR distribution for all trees sharing the same parameter
set, as Fig. 3 in Pannetier et al. (2020)
}
\author{
Théo Pannetier
}
