% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_LR_DDvTD.R
\name{plot_LR_DDvTD}
\alias{plot_LR_DDvTD}
\title{Plot logLR distribution for DD & TD trees}
\usage{
plot_LR_DDvTD(para)
}
\arguments{
\item{para}{numeric or character. A four-digits code specifying a set of parameter values. See \code{get_para_values()} for possible values.}
}
\description{
Plot logLR distribution for DD & TD trees
}
\author{
Théo Pannetier
}
