% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ltt_nested.R
\name{plot_ltt_nested}
\alias{plot_ltt_nested}
\title{Plot n distribution inside a ltt plot}
\usage{
plot_ltt_nested(para)
}
\arguments{
\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{get_para_values} doc for details, and call
\code{get_para_values} to see possible inputs.}
}
\description{
Calls \code{plot_avg_ltt} and \code{plot_n_distrib}, and nests the latter in
the former
}
\author{
Théo Pannetier
}
