% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_N_distrib_DDvTD.R
\name{plot_N_distrib_DDvTD}
\alias{plot_N_distrib_DDvTD}
\title{Plot the distribution of N for DD & TD trees}
\usage{
plot_N_distrib_DDvTD(para)
}
\arguments{
\item{para}{numeric or character. A four-digits code specifying a set of parameter values.}
}
\description{
Jointly plots the distribution of the number of tips at present for a pair 
of TD and DD trees that share the same parameter set.
}
\author{
Théo Pannetier
}
