% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_info.R
\name{plot_tree_info}
\alias{plot_tree_info}
\title{Plot main features of a simulated tree}
\usage{
plot_tree_info(sim, para)
}
\arguments{
\item{sim}{character, the simulation model.}

\item{para}{numeric or character. A four-digits code specifying a set of
parameter values.}
}
\value{
a pdf file saved at \code{DDvTD/figures/} containing a plot for
each tree. Each plot contains the plotted tree, the ltt plot, AIC score and
parameter estimates of the DD and TD models for this tree.
}
\description{
Plots a simulated tree, along with its ltt and ML estimates for both DD
and TD
}
\author{
Theo Pannetier
}
