% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_size_distribution.R
\name{plot_tree_size_distribution}
\alias{plot_tree_size_distribution}
\title{Plots the size distribution of trees in a sim dataset}
\usage{
plot_tree_size_distribution(sim)
}
\description{
Plots the nb of tips at present in the 1,000 trees of the selected tree dataset.
}
\author{
Theo Pannetier
}
