% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_estimates.R
\name{plot_trees_estimates}
\alias{plot_trees_estimates}
\title{Plot simulated trees with ML parameter estimates}
\usage{
plot_trees_estimates(sim, para, rangemc = 1:1000)
}
\arguments{
\item{sim}{name of the model used to simulate the trees. See \code{get_sim_names()}}

\item{para}{four-digit-character specifying the parameters used to simulate the trees. See \code{get_para_values}.}

\item{rangemc}{numeric vector, indexes of the trees to be plotted}
}
\description{
Plots the simulated trees in the dataset with paramter estimated with both
the DD and the TD model under each tree.
}
\author{
Théo Pannetier
}
