% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_optim_results.R
\name{read_optim_results}
\alias{read_optim_results}
\title{Read results of the model optimisation}
\usage{
read_optim_results(sim, optim, para, init_k = "true_k")
}
\arguments{
\item{sim}{character, name of the simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{optim}{character, name of the optimisation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits number coding for a set of four parameter
values. Refer to \code{arg_para()} doc for details, and call
\code{arg_para()} to see possible inputs.}

\item{init_k}{character code specifying the initial values of parameter K
used in the optimisation. Refer to \code{arg_init_k()} doc for
details, and call \code{arg_init_k()} to see possible inputs.}
}
\description{
Reads a result dataframe produced by \code{\link[=run_optim]{run_optim()}}. The file is expedcted to
be found in \verb{DDvTD/data/optim/} and have the default name computed by
\code{\link[=run_optim]{run_optim()}}.
}
\author{
Theo Pannetier
}
