% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_optim_table.R
\name{read_optim_table}
\alias{read_optim_table}
\title{Access a table with model fitting results}
\usage{
read_optim_table(sim, optim, para, init_k = "true_k")
}
\arguments{
\item{sim}{character, name of simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{optim}{character, name of the optimisation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{get_para_values} doc for details, and call
\code{get_para_values} to see possible inputs.}

\item{init_k}{character code specifying the initial values of parameter K
used in the optimisation. Refer to \code{get_possible init_k} doc for
details, and call \code{get_possible_init_k} to see possible inputs.}
}
\description{
The function will go fetch the data frame corresponding to the arguments and
return it.
}
\author{
Theo Pannetier
}
