% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sim.R
\name{read_sim}
\alias{read_sim}
\title{Load a simulated trees dataset}
\usage{
read_sim(sim, para)
}
\arguments{
\item{sim}{character, name of the simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits number coding for a set of four parameter
values. Refer to \code{arg_para()} doc for details, and call
\code{arg_para()} to see possible inputs.}
}
\description{
Loads a file containing simulated trees at the specified location.
Target file must follow nomeclature 'simXX-para.RData'
}
\author{
Theo Pannetier
}
