% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trees.R
\name{read_trees}
\alias{read_trees}
\title{Load a simulated trees dataset}
\usage{
read_trees(sim, para)
}
\arguments{
\item{sim}{character, name of simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{get_para_values} doc for details, and call
\code{get_para_values} to see possible inputs.}
}
\description{
Loads a file containing simulated trees at the specified location.
Target file must follow nomeclature 'simXX-para.RData'
}
\author{
Theo Pannetier
}
