% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_from_n.R
\name{run_from_n}
\alias{run_from_n}
\title{Run maximum likelihood esimation starting from the number of lineages}
\usage{
run_from_n(sim, optim, para, rangemc = NULL, methode = "ode45",
  optimmethod = "subplex", tol = rep(1e-06, 3), jobID = NA,
  num_cycles = 1)
}
\arguments{
\item{sim}{character, the simulation model. Call \code{get_sim_names()} for a list of possible values.}

\item{optim}{character. The name of the model to fit to the simulated phylogenies. See \code{get_optim_names()} for possible inputs.}

\item{para}{character or numeric, the code specifying the parameter values used in the simulation.
Call \code{get_para_values()} for possible inputs, and \code{para_to_pars()} for the corresponding parameter values.}

\item{rangemc}{a numeric vector containing all the indices of the trees to optimize the model on. Default to all the trees in the dataset.}

\item{methode}{argument passed to \code{\link[DDD:dd_ML]{dd_ML}} or \code{\link[DDD:dd_ML]{bd_ML}}. See the \pkg{DDD} documentation for more info.}

\item{optimmethod}{argument passed to \code{\link[DDD:dd_ML]{dd_ML}} or \code{\link[DDD:dd_ML]{bd_ML}}. See the \pkg{DDD} documentation for more info.}

\item{tol}{argument passed to \code{\link[DDD:dd_ML]{dd_ML}} or \code{\link[DDD:dd_ML]{bd_ML}}. See the \pkg{DDD} documentation for more info.}

\item{jobID}{is only relevant for metadata. Job scripts pass their ID to this argument. \code{NA} if run locally.}

\item{num_cycles}{numeric, passed to \code{dd_ML}/\code{bd_ML}. Number of cycles of optimisation.}

\item{custom_pars}{numeric vector, user-specified initial values for the parameters to be optimized.
If \code{NULL} (default), values are initialized using true values used to simulate the trees.}

\item{outputfile}{character, the name of the file to save the output data frame in.
Must be specified if \code{save_results = TRUE}.}

\item{save_results}{logical. Should save the results to \code{outputfile} (default) or not.}

\item{return_res}{logical. Should results be returned? Default to \code{FALSE}.}
}
\description{
\code{run_from_n} is simply a wrapper of \code{run_ML} where K is initialized
from the number of tips in the tree, rather than the true value.
}
\author{
Théo Pannetier
}
