% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_liklandscapes.R
\name{run_liklandscapes}
\alias{run_liklandscapes}
\title{Compute likelihood landscape along K for a range of trees}
\usage{
run_liklandscapes(sim, optim, para, fixed_pars_values = "true_values",
  outputfile = paste0("data/lik_landscapes/lik_landscape_", "s", sim,
  "_o", optim, "_", para, "_", fixed_pars_values, ".rds"))
}
\arguments{
\item{sim}{character, the name of a simulation model. See 
\code{get_sim_names()} for possible values.}

\item{optim}{character. The name of the model to get the likelihood landscape
of, can be either "DD" or "TD".}

\item{para}{numeric or character. A four-digits code specifying a set of 
parameter values. See \code{get_para_values()}}

\item{fixed_pars_values}{character specifying which values to use for lambda0
and mu0. Either 'true_values', using true values of the parameters given by
\code{para_to_pars()}; or 'ML_estimates', using maximum likelihood estimates
fetched from the ML results dataset stored in \code{data/optim/}.}

\item{outputfile}{character, the path and filename to save results at.}
}
\description{
Calls \code{get_likelihood_along_K()} for a range of trees in DDvTD
}
\author{
Théo Pannetier
}
