% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ML.R
\name{run_ML}
\alias{run_ML}
\title{Run maximum likelihood esimation of a diversification model on a set of simulated trees}
\usage{
run_ML(sim, optim, para, custom_pars = NULL, outputfile = NULL,
  rangemc = NULL, methode = "ode45", optimmethod = "subplex",
  tol = rep(1e-06, 3), save_results = TRUE, return_res = FALSE,
  jobName = NA)
}
\arguments{
\item{sim}{character, the simulation model. Call \code{get_sim_names()} for a list of possible values.}

\item{optim}{character. The name of the model to fit to the simulated phylogenies. See \code{get_optim_names()} for possible inputs.}

\item{para}{character or numeric, the code specifying the parameter values used in the simulation.
Call \code{get_para_values()} for possible inputs, and \code{para_to_pars()} for the corresponding parameter values.}

\item{custom_pars}{numeric vector, user-specified initial values for the parameters to be optimized.
If \code{NULL} (default), values are initialized using true values used to simulate the trees.}

\item{outputfile}{character, the name of the file to save the output data frame in.
Must be specified if \code{save_results = TRUE}.}

\item{rangemc}{a numeric vector containing all the indices of the trees to optimize the model on. Default to all the trees in the dataset.}

\item{methode}{argument passed to \code{\link[DDD:dd_ML]{dd_ML}} or \code{\link[DDD:dd_ML]{bd_ML}}. See the \pkg{DDD} documentation for more info.}

\item{optimmethod}{argument passed to \code{\link[DDD:dd_ML]{dd_ML}} or \code{\link[DDD:dd_ML]{bd_ML}}. See the \pkg{DDD} documentation for more info.}

\item{tol}{argument passed to \code{\link[DDD:dd_ML]{dd_ML}} or \code{\link[DDD:dd_ML]{bd_ML}}. See the \pkg{DDD} documentation for more info.}

\item{save_results}{logical. Should save the results to \code{outputfile} (default) or not.}

\item{return_res}{logical. Should results be returned? Default to \code{FALSE}.}

\item{jobID}{is only relevant for metadata. Job scripts pass their ID to this argument. \code{NA} if run locally.}
}
\description{
\code{run_ML} fetches the branching times from the dataset indicated by
\code{sim} and \code{para}, and runs the maximum loglikelihood of the model specified in \code{optim}.
}
\author{
Théo Pannetier
}
