% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_optim.R
\name{run_optim}
\alias{run_optim}
\title{Run maximum likelihood optimisation of a model on simulated trees}
\usage{
run_optim(
  sim,
  optim,
  para,
  custom_pars = NULL,
  outputfile = paste0("sim", sim, "_optim", optim, "_", para, "_true_k.rds"),
  rangemc = 1:1000,
  methode = "ode45",
  optimmethod = "subplex",
  tol = rep(1e-06, 3),
  save_results = TRUE,
  return_results = FALSE,
  jobID = NA,
  num_cycles = 1,
  cond = 1
)
}
\arguments{
\item{sim}{character, name of simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{optim}{character, name of the optimisation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{arg_para()} doc for details, and call
\code{arg_para()} to see possible inputs.}

\item{custom_pars}{numeric vector, user-specified initial values for the parameters to be optimized.
If \code{NULL} (default), values are initialized using true values used to simulate the trees.}

\item{outputfile}{character, the name of the file to save the output data
frame in. The default follows the structure expected by \link{read_optim_results}.}

\item{rangemc}{numeric vector, a set of tree indices ranging from 1 to 1000.}

\item{methode}{likelihood solving methode, passed to
\code{\link[DDD:dd_loglik]{DDD::dd_loglik()}} / \code{\link[DDD:bd_loglik]{DDD::bd_loglik()}}.}

\item{optimmethod}{optimisation algorithm, passed to
\code{\link[DDD:dd_ML]{DDD::dd_ML()}} / \code{\link[DDD:bd_ML]{DDD::bd_ML()}}.}

\item{tol}{optimisation tolerance, passed to \code{\link[DDD:dd_ML]{DDD::dd_ML()}} / \code{\link[DDD:bd_ML]{DDD::bd_ML()}}.}

\item{save_results}{logical. Should the results be saved to
\code{outputfile}?}

\item{return_results}{logical. Should the results be returned?}

\item{jobID}{SLURM job ID passed when the function is called from a
cluster script.}

\item{num_cycles}{number of cycles of optimisation,
passed to \code{\link[DDD:dd_ML]{DDD::dd_ML()}} / \code{\link[DDD:bd_ML]{DDD::bd_ML()}}.}

\item{cond}{code specifying how the likelihood is conditioned. Passed to
\code{\link[DDD:dd_loglik]{DDD::dd_loglik()}} / \code{\link[DDD:bd_loglik]{DDD::bd_loglik()}}.}
}
\description{
Read a set of simulated trees as created by \code{\link[=run_sim]{run_sim()}}, set up initial
parameter values, call \code{\link[DDD:dd_ML]{DDD::dd_ML()}} or \code{\link[DDD:bd_ML]{DDD::bd_ML()}} to run the
optimisation and format output.
}
\author{
Théo Pannetier
}
