% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_optim_empirical_sims.R
\name{run_optim_empirical_sims}
\alias{run_optim_empirical_sims}
\title{Run optimisation on trees simulated from empirical phylogenies}
\usage{
run_optim_empirical_sims(
  family_name,
  sim,
  optim,
  rangemc = 1:1000,
  outputfile = paste0(family_name, "_sim", sim, "_optim", optim, ".rds"),
  jobID = NA,
  return_results = FALSE,
  save_results = TRUE
)
}
\arguments{
\item{family_name}{character, the name of the family from which the simulated
trees were generated from. Must match a file in
\code{"DDvTD/data/empirical_trees/"}.}

\item{sim}{character, name of the simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{optim}{character, name of the optimisation model, either \code{"DD"}
or \code{"TD"}.}

\item{rangemc}{numeric vector, a set of tree indices ranging from 1 to 1000.}

\item{outputfile}{character, the name of the file to save the output data
frame in.}

\item{jobID}{SLURM job ID passed when the function is called from a
cluster script.}

\item{return_results}{logical. Should the results be returned?}

\item{save_results}{logical. Should the results be saved to
\code{outputfile}?}
}
\description{
Run optimisation on trees simulated from empirical phylogenies
}
\author{
Théo Pannetier
}
