% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_optim_from_n.R
\name{run_optim_from_n}
\alias{run_optim_from_n}
\title{Run maximum likelihood esimation starting from the number of lineages}
\usage{
run_optim_from_n(
  sim,
  optim,
  para,
  rangemc = NULL,
  methode = "ode45",
  optimmethod = "subplex",
  tol = rep(1e-06, 3),
  jobID = NA,
  num_cycles = 1,
  save_results = TRUE,
  return_res = FALSE,
  cond = 1
)
}
\arguments{
\item{sim}{character, name of simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{optim}{character, name of the optimisation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{get_para_values} doc for details, and call
\code{get_para_values} to see possible inputs.}

\item{rangemc}{numeric vector, a set of tree indices ranging from 1 to 1000.}

\item{methode}{argument passed to \code{\link[DDD:dd_ML]{dd_ML}} or
\code{\link[DDD:dd_ML]{bd_ML}}. See the \pkg{DDD} documentation for more
info.}

\item{optimmethod}{argument passed to \code{\link[DDD:dd_ML]{dd_ML}} or
\code{\link[DDD:dd_ML]{bd_ML}}. See the \pkg{DDD} documentation for more
info.}

\item{tol}{argument passed to \code{\link[DDD:dd_ML]{dd_ML}} or
\code{\link[DDD:dd_ML]{bd_ML}}. See the \pkg{DDD} documentation for more
info.}

\item{jobID}{is only relevant for metadata. Job scripts pass their ID to
this argument. \code{NA} if run locally.}

\item{num_cycles}{numeric, passed to \code{dd_ML}/\code{bd_ML}. Number of
cycles of optimisation.}

\item{save_results}{logical. Should save the results to \code{outputfile}
(default) or not.}

\item{return_res}{logical. Should results be returned? Default to
\code{FALSE}.}

\item{cond}{numeric, conditionning parameter passed to
\code{dd_ML}/\code{bd_ML}. See \code{?dd_ML} for possible values.}
}
\description{
\code{run_optim_from_n} is simply a wrapper of \code{run_optim} where K is initialized
from the number of tips in the tree, rather than the true value.
}
\author{
Théo Pannetier
}
