% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_subplex_twice.R
\name{run_subplex_twice}
\alias{run_subplex_twice}
\title{Test running optimisation twice}
\usage{
run_subplex_twice(sim, para, optim, mc, verbose = 0, methode = "ode45")
}
\arguments{
\item{sim}{character, the simulation model. Call \code{get_sim_names()} for a list of possible values.}

\item{para}{character or numeric, the code specifying the parameter values used in the simulation.
Call \code{get_para_values()} for possible inputs, and \code{para_to_pars()} for the corresponding parameter values.}

\item{optim}{character. The name of the model to fit to the simulated phylogenies. See \code{get_optim_names()} for possible inputs.}
}
\value{
a 3-element numeric vector containing the loglik after the first run, second run, and improvement
between the two (second - first)
}
\description{
A test function that calls subplex to optimise the ML, then runs it a second time
starting from the parameter estimates from the previous run.
}
\author{
Théo Pannetier
}
