% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_DDvTD_dir_struct.R
\name{set_DDvTD_dir_struct}
\alias{set_DDvTD_dir_struct}
\title{Set the structure of the DDvTD directory and enter it}
\usage{
set_DDvTD_dir_struct()
}
\description{
Creates a "DDvTD" folder and the nested folders used by functions in
\code{DDvTDtools}.
}
\author{
Theo Pannetier
}
