% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trees.R
\name{simulate_trees}
\alias{simulate_trees}
\title{Simulate a set of phylogenetic trees under a time-dependent or
diversity-dependent model of diversification.}
\usage{
simulate_trees(
  sim,
  para = NULL,
  custom_pars = NULL,
  nbmc = 1000,
  outputfile = paste0("./data/sim/sim", sim, "-", para, ".RData"),
  save_results = TRUE,
  return_results = TRUE
)
}
\arguments{
\item{sim}{a character used to select the model, can be either 'DD' or
'TD'.}

\item{para}{a 4-digit code used to set the parameter values used for the
simulation. See \code{parameter_list} for possible values.}

\item{custom_pars}{a 4-element numeric vector containing c(age, pars)
to supply custom parameters (i.e., values not covered with the \code{para}
option). Either \code{para} or \code{custom_pars} must be supplied, not both.}

\item{nbmc}{Integer. How many trees should be generated?}

\item{outputfile}{path and name for output file, by default generated
automatically from \code{sim} and \code{para}.}

\item{save_results}{logical. Should save the results to \code{outpufile}
(default) or not.}

\item{return_results}{logical. Should results be returned? Default to
\code{TRUE}.}
}
\value{
A \code{list} containing itself \code{nbmc} lists corresponding a
simulated trees. See \code{DDD:dd_sim} for the content of these lists.
}
\description{
Reads a model name and coded parameter input to simulate a dataset of a
number of simulated phylogenetic trees.
}
\author{
Cesar Martinez and Theo Pannetier
}
