% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_para.R
\name{arg_para}
\alias{arg_para}
\title{Get possible values of para argument}
\usage{
arg_para(incl_unused = FALSE)
}
\arguments{
\item{incl_unused}{logical, specify whether you want to include values that
have been considered in earlier phases of the project.}
}
\description{
Input values currently accepted for para.
Use \code{para_to_pars} to translate para into paramter values.
}
\seealso{
\code{para_to_pars()}
}
\author{
Theo Pannetier
}
