% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_landscape_data.R
\name{assemble_landscape_data}
\alias{assemble_landscape_data}
\title{Assemble log-likelihood landscape data for plotting}
\usage{
assemble_landscape_data(sim, para, mc)
}
\arguments{
\item{sim}{character, name of simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{get_para_values} doc for details, and call
\code{get_para_values} to see possible inputs.}

\item{mc}{numeric between 1 and 1000, the index of a tree within a dataset.}
}
\description{
Fetches saved results for likelihood landscapes along K and assembles a data
frame ready for plotting
}
\author{
Théo Pannetier
}
