% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_init_k.R
\name{assert_init_k}
\alias{assert_init_k}
\title{Asserts input init parameter is supported}
\usage{
assert_init_k(init_k)
}
\arguments{
\item{init_k}{character code specifying the initial values of parameter K
used in the optimisation. Refer to \code{get_possible init_k()} doc for
details, and call \code{arg_init_k()} to see possible inputs.}
}
\description{
Asserts input init parameter is supported
}
\author{
Theo Pannetier
}
