% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_optim.R
\name{assert_optim}
\alias{assert_optim}
\title{Assert the value of 'optim' is a valid model name for DDvTDtools}
\usage{
assert_optim(optim)
}
\arguments{
\item{optim}{character. The name of a diversification model fitted to data.}
}
\description{
Assert the value of 'optim' is a valid model name for DDvTDtools
}
\author{
Theo Pannetier
}
