% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_para.R
\name{assert_para}
\alias{assert_para}
\title{Assert the value of 'para' is a valid parameter combination for DDvTDtools}
\usage{
assert_para(para)
}
\arguments{
\item{para}{numeric, a four-digits number coding for a set of four parameter
values. Refer to \code{arg_para()} doc for details, and call
\code{arg_para()} to see possible inputs.}
}
\description{
Assert the value of 'para' is a valid parameter combination for DDvTDtools
}
\seealso{
\code{arg_para()}, \code{para_to_pars()}
}
\author{
Theo Pannetier
}
