% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convergence_optim.R
\name{check_convergence_optim}
\alias{check_convergence_optim}
\title{Check optimisation results for completeness}
\usage{
check_convergence_optim(folder = "data/optim/")
}
\arguments{
\item{folder}{the folder where the results files are stored. All .rds files in the folder are read.}
}
\description{
Reads result data frames in the optim folder and prints a message if there is
any missing row.
}
\author{
Théo Pannetier
}
