% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_results.R
\name{check_incomplete_rows}
\alias{check_incomplete_rows}
\title{Check a results data frame for NAs in rows}
\usage{
check_incomplete_rows(results, return_rows = TRUE)
}
\arguments{
\item{results}{a results data frame, as produced by \code{\link[=run_optim]{run_optim()}} or accessed by
\code{\link[=read_optim_results]{read_optim_results()}}.}

\item{return_rows}{logical. Should rows that don't pass the check be
returned?}
}
\description{
Goes through rows and returns a message if any NA is found in  row.
}
\author{
Theo Pannetier
}
