% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_results.R
\name{check_missing_mcs}
\alias{check_missing_mcs}
\title{Check a resultsults data frame for missing rows}
\usage{
check_missing_mcs(results, return_rows)
}
\arguments{
\item{results}{a resultsults data frame, as produced by [run_optim()] or accessed by
[read_optim_resultsults].}

\item{return_rows}{logical. Should rows that don't pass the check be
returned?}
}
\description{
Goes through rows and returns a message if any row from 1 to 1000 is missing
}
\author{
Theo Pannetier
}
