% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_errors.R
\name{count_errors}
\alias{count_errors}
\title{Count errors in a log file}
\usage{
count_errors(log_filename)
}
\arguments{
\item{log_file}{a log file to search the errors in. Can be a \code{.txt} or
\code{.log} file.}
}
\value{
a list with the indexes of trees that returned each error.
Also prints the nb of occurrences of each error to the console
}
\description{
Counts occcurrences of common errors in log files generated by
\code{run_ML()}.
}
\author{
Théo Pannetier
}
