% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_pars2_dd1.R
\name{default_pars2_dd1}
\alias{default_pars2_dd1}
\title{Predefined pars2 argument for dd_loglik as used in DDvTD}
\usage{
default_pars2_dd1()
}
\description{
Returns default settings to call dd_loglik() for DDvTD, model DD1
}
\author{
Theo Pannetier
}
