% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_pars2_td4.R
\name{default_pars2_td4}
\alias{default_pars2_td4}
\title{Predefined pars2 argument for bd_loglik as used in DDvTD}
\usage{
default_pars2_td4()
}
\description{
Returns default settings to call bd_loglik() for DDvTD, model TD4
}
\author{
Theo Pannetier
}
