% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_init_k.R
\name{fetch_init_k}
\alias{fetch_init_k}
\title{Fetch values for argument init_k}
\usage{
fetch_init_k()
}
\description{
Labels for initial value of K associated with each parameter setting for the
main analysis.
}
\author{
Théo Pannetier
}
