% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_rows.R
\name{fill_missing_rows}
\alias{fill_missing_rows}
\title{Fill missing results row with default values}
\usage{
fill_missing_rows(df, para)
}
\arguments{
\item{df}{a results data frame, i.e. as read by
\code{DDvTDtools::read_optim_table()}}

\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{get_para_values} doc for details, and call
\code{get_para_values} to see possible inputs.}
}
\description{
Function used to fill result data for trees that miss any result.
Fills the result matrix for a parameter set with default values for these
missing trees
}
\author{
Théo Pannetier
}
