% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_rows.R
\name{fill_missing_rows}
\alias{fill_missing_rows}
\title{Fill missing results row with default values}
\usage{
fill_missing_rows(results, para)
}
\arguments{
\item{results}{a results data frame, as produced by \code{\link[=run_optim]{run_optim()}} or accessed
through \code{\link[=read_optim_results]{read_optim_results()}}.}

\item{para}{numeric, a four-digits number coding for a set of four parameter
values. Refer to \code{arg_para()} doc for details, and call
\code{arg_para()} to see possible inputs.}
}
\description{
Searches a results data frame for missing rows and fills them with default
values.
}
\author{
Théo Pannetier
}
