% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_indexing.R
\name{fix_indexing}
\alias{fix_indexing}
\title{Fix indexing of results data frame}
\usage{
fix_indexing(results)
}
\arguments{
\item{results}{a results data frame, as produced by [run_optim()] or accessed
through [read_optim_results()].}
}
\description{
Checks if rows in the results data frame are indexed correctly
(from 1 to 1000), and, if not, re-orders them accordingly.
}
\author{
Theo Pannetier
}
