% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_avg_ltt_table.R
\name{get_avg_ltt_table}
\alias{get_avg_ltt_table}
\title{Computes the average number of lineages over time for a set of phylogenies}
\usage{
get_avg_ltt_table(sim, para, with_extinct = FALSE)
}
\arguments{
\item{sim}{character, name of simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{get_para_values} doc for details, and call
\code{get_para_values} to see possible inputs.}

\item{with_extinct}{logical, should the tree include extinct lineages?}
}
\description{
Returns the average number of lineage over time for a dataset of 1000 simulated
phylogenies
}
\author{
Théo Pannetier
}
