% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_avg_ltt_table.R
\name{get_avg_ltt_table}
\alias{get_avg_ltt_table}
\title{Computes the average number of lineages over time for a set of phylogenies}
\usage{
get_avg_ltt_table(sim, para, with_extinct = FALSE)
}
\arguments{
\item{sim}{character, name of the simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits number coding for a set of four parameter
values. Refer to \code{arg_para()} doc for details, and call
\code{arg_para()} to see possible inputs.}

\item{with_extinct}{logical, should the tree include extinct lineages?}
}
\description{
Computes a table containing the average number of species over time over
1000 replicates of a simulated set of trees. The averaging is performed with
\code{\link[TreeSim:LTT.plot.gen]{TreeSim::LTT.plot.gen()}}.
}
\author{
Théo Pannetier
}
