% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brts.R
\name{get_brts}
\alias{get_brts}
\title{Extract a set of branching times from a dataset of simulated trees}
\usage{
get_brts(sim, para, mc, with_extinct = F)
}
\arguments{
\item{sim}{character, the name of a simulation model. See
\code{get_sim_names()} for possible values.}

\item{para}{numeric or character. A four-digits code specifying a set of
parameter values. See \code{get_para_values()}}

\item{mc}{numeric. The index of the tree to return branching times from.}

\item{with_extinct}{logical. \code{with_extinct = FALSE} for the
reconstructed tree, \code{with_extinct = TRUE} for the complete tree.}
}
\value{
a numerical vector containing the branching times of the tree.
}
\description{
Read the dataset and returns the branching times for the specified tree.
}
\author{
Theo Pannetier
}
