% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dAIC_DDvTD.R
\name{get_dAIC_DDvTD}
\alias{get_dAIC_DDvTD}
\title{Compute AIC_DD - AIC_TD for a set of trees}
\usage{
get_dAIC_DDvTD(sim, para)
}
\arguments{
\item{sim}{character, name of simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{get_para_values} doc for details, and call
\code{get_para_values} to see possible inputs.}
}
\description{
Compute AIC_DD - AIC_TD for a set of trees
}
\author{
Théo Pannetier
}
