% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_initpars.R
\name{get_default_initpars}
\alias{get_default_initpars}
\title{Get default initial parameter values for run_optim optimisation}
\usage{
get_default_initpars(true_pars, optim, brts)
}
\arguments{
\item{true_pars}{numeric vector. The true parameter values used to simulate
the trees.}

\item{optim}{character, name of the optimisation model, either \code{"DD"}
or \code{"TD"}.}

\item{brts}{numeric vector, a set of branching times as given by
\code{ape::branching_times()}.}
}
\description{
Returns default initial parameter values used to fit \code{optim} models to
phylogenies with \code{run_optim()}. Called by \code{run_optim()}.
}
\details{
For the DD and TD models, initial parameter values are by default
set to the true values used to simulate the trees. For thr CR model, la0
is set to the expected speciation rate, knowing the age, number of tips and
the extinction rate for the tree. This is computed using equation (15) from
Kendall (1948) for the expected number of particles over time in a
constant-rate birth-death process.
}
\references{
Kendall, David G. On the Generalized "Birth-and-Death"
Process. Ann. Math. Statist. 19 (1948), no. 1, 1--15.
doi:10.1214/aoms/1177730285
}
\author{
Théo Pannetier
}
