% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_likelihood_along_K.R
\name{get_likelihood_along_K}
\alias{get_likelihood_along_K}
\title{Compute the likelihood landscape along the K-axis}
\usage{
get_likelihood_along_K(optim, brts, lambda0, mu0, K_seq = seq(10, 1000,
  by = 10))
}
\arguments{
\item{optim}{character. The name of the model to fit to the simulated
phylogenies, either "DD" or "TD".}

\item{brts}{numeric vector, a set of branching times.}

\item{lambda0}{numeric. A (fixed) value for the speciation rate.}

\item{mu0}{numeric. A (fixed) value for the extinction rate.}

\item{K_seq}{numeric vector, the values of parameter K to return values for.}
}
\description{
Returns the sequence of (log-)likelihood estimates of a tree under DD or TD
for a vector of values of parameter K.
}
\details{
Values of the two other parameters of the model (\code{lamba0} and
\code{mu0}) are taken from the maximum likelihood estimate. These values are
fetched from the results data frame with \code{read_optim_table}, so you need
to make sure they exist and that the file matches the package syntax to run
the function.
}
\author{
Théo Pannetier
}
