% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lr.R
\name{get_lr}
\alias{get_lr}
\title{Compute likelihood ratios for a results dataset}
\usage{
get_lr(sim, para, init_k = "true_k")
}
\arguments{
\item{sim}{character, name of the simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits number coding for a set of four parameter
values. Refer to \code{arg_para()} doc for details, and call
\code{arg_para()} to see possible inputs.}

\item{init_k}{character code specifying the initial values of parameter K
used in the optimisation. Refer to \code{arg_init_k()} doc for
details, and call \code{arg_init_k()} to see possible inputs.}
}
\description{
Compute likelihood ratios for a results dataset
}
\author{
Théo Pannetier
}
