% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lr_table.R
\name{get_lr_table}
\alias{get_lr_table}
\title{Assemble a table containing likelihood ratios}
\usage{
get_lr_table(para, init_k)
}
\arguments{
\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{get_para_values} doc for details, and call
\code{get_para_values} to see possible inputs.}

\item{init_k}{character code specifying the initial values of parameter K
used in the optimisation. Refer to \code{get_possible init_k} doc for
details, and call \code{get_possible_init_k} to see possible inputs.}
}
\description{
Calls \code{get_lr} and binds results in a table with corresponding
\code{sim} ans \code{para} values.
}
\author{
Theo Pannetier
}
