% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lr_threshold.R
\name{get_lr_threshold}
\alias{get_lr_threshold}
\title{Get the relevant log LR threshold value to reject either model}
\usage{
get_lr_threshold(sim, para, init_k, percentile)
}
\arguments{
\item{percentile}{numeric value between 0 and 1, the percentile on which the
threshold should be based (normally 0.05 for DD, 0.95 for TD).}
}
\description{
Computes the threshold value for the log-likelihood ratio to reject either
model.
}
\author{
Theo Pannetier
}
