% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multi_brts.R
\name{get_multi_brts}
\alias{get_multi_brts}
\title{Extract multiple sets of branching times from a dataset of simulated trees}
\usage{
get_multi_brts(dir, sim, para, rangemc, with_extinct = F)
}
\arguments{
\item{dir}{character, the directory where simulated datasets are stored}

\item{sim}{character, the name of a simulation model. See \code{get_sim_names()} for possible values.}

\item{para}{numeric or character. A four-digits code specifying a set of parameter values. See \code{get_para_values()}}

\item{rangemc}{numeric vector. The indices of the trees to return branching times from.}

\item{with_extinct}{logical. \code{with_extinct = F} for the reconstructed tree, \code{with_extinct = T} for the complete tree.}
}
\value{
a list of numerical vectors containing the branching times of the specified trees.
}
\description{
Read the dataset at specified location and returns branching times for multiple trees specified by indices.
}
\author{
Theo Pannetier
}
