% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_table.R
\name{get_n_table}
\alias{get_n_table}
\title{Get N for each tree in a dataset}
\usage{
get_n_table(sim, para)
}
\arguments{
\item{para}{}
}
\description{
Assemble a table containing the number of tips in each tree in a results data
set.
}
\author{
Theo Pannetier
}
