% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optim_df_row.R
\name{get_optim_df_row}
\alias{get_optim_df_row}
\title{Format run_ML optimisation results in a data frame row}
\usage{
get_optim_df_row(mc, sim, optim, brts, true_pars, init_pars, ML_output,
  num_cycles = 1, methode, optimmethod, jobID)
}
\description{
Internal function called only by \code{run_ML}.
}
\author{
Théo Pannetier
}
