% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_success.R
\name{get_p_success}
\alias{get_p_success}
\title{Computes the proportion of trees that can successfully be assigned to either
model}
\usage{
get_p_success(sim, para, init_k)
}
\arguments{
\item{sim}{character, name of simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{arg_para()} doc for details, and call
\code{arg_para()} to see possible inputs.}

\item{init_k}{character code specifying the initial values of parameter K
used in the optimisation. Refer to \code{get_possible init_k()} doc for
details, and call \code{arg_init_k()} to see possible inputs.}
}
\description{
Computes the proportion of trees in the dataset that reject the opposite
model, i.e. logLR is beyond the corresponding threshold computed by
\code{\link[=get_lr_threshold]{get_lr_threshold()}}.
}
\author{
Theo Pannetier
}
