% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_para_values.R
\name{get_para_values}
\alias{get_para_values}
\title{Get possible values of para argument}
\usage{
get_para_values()
}
\description{
Input values currently accepted for para.
   Use \code{para_to_pars} to translate para into paramter values.
}
\seealso{
\code{para_to_pars()}
}
\author{
Theo Pannetier
}
