% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_multiPhylo.R
\name{get_sim_multiPhylo}
\alias{get_sim_multiPhylo}
\title{Load a simulated dataset as a multiPhylo object}
\usage{
get_sim_multiPhylo(sim, para, with_extinct = F)
}
\arguments{
\item{sim}{name of the model used to simulate the trees. See \code{get_sim_names()}}

\item{para}{four-digit-character specifying the parameters used to simulate the trees. See \code{get_para_values}.}
}
\description{
Load a simulated dataset as a multiPhylo object
}
\author{
Theo Pannetier
}
