% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_multiPhylo.R
\name{get_sim_multiPhylo}
\alias{get_sim_multiPhylo}
\title{Load a simulated dataset as a multiPhylo object}
\usage{
get_sim_multiPhylo(sim, para, with_extinct = F)
}
\arguments{
\item{sim}{character, name of simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{get_para_values} doc for details, and call
\code{get_para_values} to see possible inputs.}

\item{with_extinct}{logical, should the tree include extinct lineages?}
}
\description{
Load a simulated dataset as a multiPhylo object
}
\author{
Theo Pannetier
}
