% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para_to_pars.R
\name{para_to_pars}
\alias{para_to_pars}
\title{Translate 'para' argument into parameter values}
\usage{
para_to_pars(para)
}
\arguments{
\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{get_para_values} doc for details, and call
\code{get_para_values} to see possible inputs.}
}
\value{
a four-element numeric vector containing the parameter values associated to 'para'.
}
\description{
Prints the parameter values associated with a given value of argument 'para'.
}
\author{
Theo Pannetier
}
