% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_avg_ltt.R
\name{plot_avg_ltt}
\alias{plot_avg_ltt}
\title{Plot average lineage-through-time plots}
\usage{
plot_avg_ltt(para, with_extinct = FALSE, trees = NULL, alpha = 0.05)
}
\arguments{
\item{para}{numeric, a four-digits number coding for a set of four parameter
values. Refer to \code{arg_para()} doc for details, and call
\code{arg_para()} to see possible inputs.}

\item{with_extinct}{logical, should the tree include extinct lineages?}

\item{trees}{a vector of trees indices. If supplied, the LTTs of these trees
will also be plotted.}

\item{alpha}{transparence level for LTTs, if \code{trees} is supplied.}
}
\description{
Plots the lineage-through-time curves for a pair of sets of
(diversity-dependent and time-dependent) simulated trees. Each curve is
averaged over 1000 trees with \code{\link[TreeSim:LTT.plot.gen]{TreeSim::LTT.plot.gen()}}.
}
\author{
Théo Pannetier
}
