% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_avg_ltt.R
\name{plot_avg_ltt}
\alias{plot_avg_ltt}
\title{Plot the average LTTs of paired DD and TD trees}
\usage{
plot_avg_ltt(para, with_extinct = FALSE)
}
\arguments{
\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{arg_para()} doc for details, and call
\code{arg_para()} to see possible inputs.}

\item{with_extinct}{logical, should the tree include extinct lineages?}
}
\description{
Plots the average LTTs of a set of DD and TD trees that share the same parameter set.
}
\author{
Théo Pannetier
}
