% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_blrt.R
\name{plot_blrt}
\alias{plot_blrt}
\title{Plot logLR distribution for DD & TD trees}
\usage{
plot_blrt(
  para,
  init_k = "true_k",
  plot_thresholds = TRUE,
  label_success = TRUE,
  threshold_dd = 0.05,
  threshold_td = 0.95,
  full_range = FALSE,
  which_geom = "density"
)
}
\arguments{
\item{para}{numeric or character. A four-digits code specifying a set of
parameter values. See \code{get_para_values()} for possible values.}

\item{init_k}{character, the setting used to initialise parameter K.
See \code{get_possible_init_k()} for possible values.}

\item{plot_thresholds}{logical, should quantile bars be plotted with the
distribution?}

\item{label_success}{logical, should the proportion of successes be added as
a label on the plot?}

\item{threshold_dd}{numeric, quantiles of the lr distribution that should be
added to the plot.}

\item{threshold_td}{numeric, quantiles of the lr distribution that should be
added to the plot.}

\item{full_range}{logical. If \code{TRUE}, the range of the x-axis is set to
include all points in the dataset. Otherwise, the default is
\code{c(-10, 10)}.}

\item{which_geom}{character giving the type of plot to be drawn. Options are
\code{"density"} and \code{"histograms"}.}
}
\description{
Plot logLR distribution for DD & TD trees
}
\author{
Théo Pannetier
}
