% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loglik_landscape_K.R
\name{plot_loglik_landscape_K}
\alias{plot_loglik_landscape_K}
\title{Plot likelihood landscapes along K}
\usage{
plot_loglik_landscape_K(sim, para)
}
\arguments{
\item{sim}{character, the name of a simulation model. See
\code{get_sim_names()} for possible values.}

\item{para}{numeric or character. A four-digits code specifying a set of
parameter values.}
}
\description{
Plot likelihood curves of DD and TD models over parameter K from saved
results.
}
\author{
Théo Pannetier
}
