% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_LR_DDvTD.R
\name{plot_LR_DDvTD}
\alias{plot_LR_DDvTD}
\title{Plot logLR distribution for DD & TD trees}
\usage{
plot_LR_DDvTD(para, init_k = "true_k", add_quantiles = FALSE,
  quantiles_dd = c(0, 0.95), quantiles_td = c(0, 0.95),
  which_geom = "histogram")
}
\arguments{
\item{para}{numeric or character. A four-digits code specifying a set of
parameter values. See \code{get_para_values()} for possible values.}

\item{init_k}{character, the setting used to initialise parameter K.
See \code{get_possible_init_k()} for possible values.}

\item{add_quantiles}{logical, should quantile bars be plotted with the
distribution ?}

\item{quantiles_dd}{numeric, quantiles of the lr distribution that should be
added to the plot.}

\item{quantiles_td}{numeric, quantiles of the lr distribution that should be
added to the plot.}

\item{which_geom}{character giving the type of plot to be drawn. Options are
\code{"density"} and \code{"histograms"}.}
}
\description{
Plot logLR distribution for DD & TD trees
}
\author{
Théo Pannetier
}
