% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ltt_nested.R
\name{plot_ltt_nested}
\alias{plot_ltt_nested}
\title{Plot n distribution inside a ltt plot}
\usage{
plot_ltt_nested(para)
}
\arguments{
\item{para}{numeric or character. A four-digits code specifying a set of
parameter values.}
}
\description{
Calls \code{plot_avg_ltt} and \code{plot_n_distrib}, and nests the latter in
the former
}
\author{
Théo Pannetier
}
