% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ltt_nested.R
\name{plot_ltt_nested}
\alias{plot_ltt_nested}
\title{Plot n distribution inside a ltt plot}
\usage{
plot_ltt_nested(para)
}
\arguments{
\item{para}{numeric, a four-digits number coding for a set of four parameter
values. Refer to \code{arg_para()} doc for details, and call
\code{arg_para()} to see possible inputs.}
}
\description{
Plots \code{\link[=plot_n_distrib]{plot_n_distrib()}} inside \code{\link[=plot_avg_ltt]{plot_avg_ltt()}}, as Fig. 2 in
Pannetier et al. (2020).
}
\author{
Théo Pannetier
}
