% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantiles.R
\name{plot_quantiles}
\alias{plot_quantiles}
\title{Add quantile bars to logLR distribution plot}
\usage{
plot_quantiles(df, which_dd = 0.05, which_td = 0.95)
}
\arguments{
\item{df}{data frame containing the data, as passed from \code{plot_LR_DDvTD}}

\item{quantile_dd}{numeric,}
}
\description{
Extracts quantiles from the loglik ratios distribution and add them to the
plot
}
