% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_optim_table.R
\name{read_optim_table}
\alias{read_optim_table}
\title{Access a table with model fitting results}
\usage{
read_optim_table(dir, sim, optim, init, para)
}
\arguments{
\item{dir}{character, path to the directory where the target file is located.}

\item{sim}{name of the model used to simulate the trees. See \code{get_sim_names()}}

\item{optim}{name of the model fitted to the simulated trees. See \code{get_optim_names()}}

\item{init}{integer, the initialization parameter of model fitting.}

\item{para}{four-digit-character specifying the parameters used to simulate the trees. See \code{get_para_values}.}
}
\description{
The function will go fetch the data frame corresponding to the arguments and return it.
}
\author{
Theo Pannetier
}
