% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_optim_struct.R
\name{results_optim_struct}
\alias{results_optim_struct}
\title{Get an empty data frame to initiate run_optim output.}
\usage{
results_optim_struct()
}
\description{
Internal function called only by \code{\link[=run_optim]{run_optim()}}.
}
\author{
Théo Pannetier
}
