% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_liklandscapes.R
\name{run_liklandscapes}
\alias{run_liklandscapes}
\title{Compute likelihood landscape along K for a range of trees}
\usage{
run_liklandscapes(
  sim,
  optim,
  para,
  rangemc = seq(1, 1000, by = 50),
  fixed_pars_values = "true_values",
  K_seq = NULL,
  outputfile = paste0("data/lik_landscapes/lik_landscape_", "s", sim, "_o", optim, "_",
    para, "_", fixed_pars_values, ".rds")
)
}
\arguments{
\item{sim}{character, name of simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{optim}{character, name of the optimisation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{get_para_values} doc for details, and call
\code{get_para_values} to see possible inputs.}

\item{rangemc}{numeric vector, a set of tree indices ranging from 1 to 1000.}

\item{fixed_pars_values}{character specifying which values to use for lambda0
and mu0. 'true_values', to use true values of the parameters as given by
\code{para_to_pars()}; 'true_k', to fetch ML estimates obtained starting
from true K, or 'from_n', to fetch ML estimates obtained starting from N.}

\item{K_seq}{numeric vector, the values of K to get the likelihood for, i.e.
the points on which the landscape will be drawn. Default option is every 10
up to 100, and every 100 up to 1000, plus increased resolution around both N
and the true value of K (40).}

\item{outputfile}{character, the path and filename to save results at.}
}
\description{
Calls \code{get_likelihood_along_K()} for a range of trees in DDvTD
}
\author{
Théo Pannetier
}
