% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_optim.R
\name{run_optim}
\alias{run_optim}
\title{Run maximum likelihood esimation of a diversification model on a set of simulated trees}
\usage{
run_optim(
  sim,
  optim,
  para,
  custom_pars = NULL,
  outputfile = NULL,
  rangemc = NULL,
  methode = "ode45",
  optimmethod = "subplex",
  tol = rep(1e-06, 3),
  save_results = TRUE,
  return_res = FALSE,
  jobID = NA,
  num_cycles = 1,
  cond = 1
)
}
\arguments{
\item{sim}{character, name of simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{optim}{character, name of the optimisation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{get_para_values} doc for details, and call
\code{get_para_values} to see possible inputs.}

\item{custom_pars}{numeric vector, user-specified initial values for the parameters to be optimized.
If \code{NULL} (default), values are initialized using true values used to simulate the trees.}

\item{outputfile}{character, the name of the file to save the output data frame in.
Must be specified if \code{save_results = TRUE}.}

\item{rangemc}{numeric vector, a set of tree indices ranging from 1 to 1000.}

\item{methode}{argument passed to \code{\link[DDD:dd_ML]{dd_ML}} or \code{\link[DDD:dd_ML]{bd_ML}}. See the \pkg{DDD} documentation for more info.}

\item{optimmethod}{argument passed to \code{\link[DDD:dd_ML]{dd_ML}} or \code{\link[DDD:dd_ML]{bd_ML}}. See the \pkg{DDD} documentation for more info.}

\item{tol}{argument passed to \code{\link[DDD:dd_ML]{dd_ML}} or \code{\link[DDD:dd_ML]{bd_ML}}. See the \pkg{DDD} documentation for more info.}

\item{save_results}{logical. Should save the results to \code{outputfile} (default) or not.}

\item{return_res}{logical. Should results be returned? Default to \code{FALSE}.}

\item{jobID}{is only relevant for metadata. Job scripts pass their ID to this argument. \code{NA} if run locally.}

\item{num_cycles}{numeric, passed to \code{dd_ML}/\code{bd_ML}. Number of cycles of optimisation.}

\item{cond}{numeric, conditionning parameter passed to
\code{dd_ML}/\code{bd_ML}. See \code{?dd_ML} for possible values.}
}
\description{
\code{run_optim} fetches the branching times from the dataset indicated by
\code{sim} and \code{para}, and runs the maximum loglikelihood of the model specified in \code{optim}.
}
\author{
Théo Pannetier
}
