% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_optim_from_n.R
\name{run_optim_from_n}
\alias{run_optim_from_n}
\title{Run maximum likelihood esimation starting from the number of lineages}
\usage{
run_optim_from_n(
  sim,
  optim,
  para,
  rangemc = 1:1000,
  methode = "ode45",
  optimmethod = "subplex",
  tol = rep(1e-06, 3),
  jobID = NA,
  num_cycles = 1,
  save_results = TRUE,
  return_results = FALSE,
  cond = 1
)
}
\arguments{
\item{sim}{character, name of simulation model, either \code{"DD"}
or \code{"TD"}.}

\item{optim}{character, name of the optimisation model, either \code{"DD"}
or \code{"TD"}.}

\item{para}{numeric, a four-digits code specifying a set of four parameter
values. Refer to \code{arg_para()} doc for details, and call
\code{arg_para()} to see possible inputs.}

\item{rangemc}{numeric vector, a set of tree indices ranging from 1 to 1000.}

\item{methode}{likelihood solving methode, passed to
[DDD::dd_ML()] / [DDD::bd_ML()].}

\item{optimmethod}{optimisation algorithm, passed to
[DDD::dd_ML()] / [DDD::bd_ML()].}

\item{tol}{optimisation tolerance, passed to [DDD::dd_ML()] / [DDD::bd_ML()].}

\item{jobID}{SLURM job ID passed when the function is called from a
cluster script.}

\item{num_cycles}{number of cycles of optimisation,
passed to [DDD::dd_ML()] / [DDD::bd_ML()].}

\item{save_results}{logical. Should the results be saved to
\code{outputfile}?}

\item{return_results}{logical. Should the results be returned?}

\item{cond}{code specifying how the likelihood is conditioned. Passed to
[DDD::dd_ML()] / [DDD::bd_ML()].}
}
\description{
An alternative version of [run_optim()] where the initial value of K is set
to N, the nb of tips in the trees, in order to ease convergence to the
maximum likelihood.
}
\author{
Théo Pannetier
}
